+++ src/Datetime.h
@@ -134,7 +134,7 @@ class Datetime
 
 	static long sTimeToMilliSecs(string sTime);
 	static time_t parseUtcStringToUtcInSecs(const string &datetime);
-	static long long parseUtcStringToUtcInMillisecs(const string &datetime);
+	static int64_t parseUtcStringToUtcInMillisecs(const string &datetime);
 	static int64_t sDateMilliSecondsToUtc(string sDate);
 	static string utcToUtcString(time_t utc);
 	static string utcToLocalString(time_t utc);
